+++ packages/react/fast-components-react-msft/src/context-menu/context-menu.props.ts
@@ -1,5 +1,4 @@
 import React from "react";
-import { Subtract } from "utility-types";
 import {
     ContextMenuHandledProps as BaseContextMenuHandledProps,
     ContextMenuManagedClasses as BaseContextMenuManagedClasses,
@@ -13,7 +12,7 @@ import {
 export type ContextMenuManagedClasses = ManagedClasses<ContextMenuClassNameContract>;
 export interface ContextMenuHandledProps
     extends ContextMenuManagedClasses,
-        Subtract<BaseContextMenuHandledProps, BaseContextMenuManagedClasses> {}
+        Omit<BaseContextMenuHandledProps, keyof BaseContextMenuManagedClasses> {}
 
 export type ContextMenuUnhandledProps = BaseContextMenuUnhandledProps;
 export type ContextMenuProps = ContextMenuHandledProps & ContextMenuUnhandledProps;