+++ packages/react/fast-components-react-msft/src/label/index.ts
@@ -11,7 +11,6 @@ import {
 } from "@microsoft/fast-components-react-base";
 import manageJss, { ManagedJSSProps } from "@microsoft/fast-jss-manager-react";
 import { DesignSystem, LabelStyles } from "@microsoft/fast-components-styles-msft";
-import { Subtract } from "utility-types";
 import labelSchema from "./label.schema";
 import labelSchema2 from "./label.schema.2";
 
@@ -22,7 +21,7 @@ import labelSchema2 from "./label.schema.2";
 const Label = manageJss(LabelStyles)(BaseLabel);
 type Label = InstanceType<typeof Label>;
 
-type LabelHandledProps = Subtract<BaseLabelHandledProps, LabelManagedClasses>;
+type LabelHandledProps = Omit<BaseLabelHandledProps, keyof LabelManagedClasses>;
 type LabelProps = ManagedJSSProps<BaseLabelProps, LabelClassNameContract, DesignSystem>;
 
 export {