+++ packages/react/fast-components-react-msft/src/pivot/index.ts
@@ -2,7 +2,6 @@ import React from "react";
 import { PivotClassNameContract } from "@microsoft/fast-components-class-name-contracts-msft";
 import manageJss, { ManagedJSSProps } from "@microsoft/fast-jss-manager-react";
 import { DesignSystem, PivotStyles } from "@microsoft/fast-components-styles-msft";
-import { Subtract } from "utility-types";
 import pivotSchema from "./pivot.schema";
 import pivotSchema2 from "./pivot.schema.2";
 import MSFTPivot, {
@@ -15,7 +14,7 @@ import MSFTPivot, {
 const Pivot = manageJss(PivotStyles)(MSFTPivot);
 type Pivot = InstanceType<typeof Pivot>;
 
-type PivotHandledProps = Subtract<MSFTPivotHandledProps, PivotManagedClasses>;
+type PivotHandledProps = Omit<MSFTPivotHandledProps, keyof PivotManagedClasses>;
 type PivotProps = ManagedJSSProps<MSFTPivotProps, PivotClassNameContract, DesignSystem>;
 
 export {