+++ packages/react/fast-components-react-msft/src/slider-label/index.ts
@@ -2,7 +2,6 @@ import React from "react";
 import manageJss, { ManagedJSSProps } from "@microsoft/fast-jss-manager-react";
 import { SliderLabelClassNameContract } from "@microsoft/fast-components-class-name-contracts-msft";
 import { DesignSystem, SliderLabelStyles } from "@microsoft/fast-components-styles-msft";
-import { Subtract } from "utility-types";
 import sliderLabelSchema from "./slider-label.schema";
 import sliderLabelSchema2 from "./slider-label.schema.2";
 import MSFTSliderLabel, {
@@ -15,9 +14,9 @@ import MSFTSliderLabel, {
 const SliderLabel = manageJss(SliderLabelStyles)(MSFTSliderLabel);
 type SliderLabel = InstanceType<typeof SliderLabel>;
 
-type SliderLabelHandledProps = Subtract<
+type SliderLabelHandledProps = Omit<
     MSFTSliderLabelHandledProps,
-    SliderLabelManagedClasses
+    keyof SliderLabelManagedClasses
 >;
 type SliderLabelProps = ManagedJSSProps<
     MSFTSliderLabelProps,