+++ packages/react/fast-components-react-msft/src/subheading/index.ts
@@ -3,7 +3,6 @@ import { FoundationProps } from "@microsoft/fast-components-foundation-react";
 import { SubheadingClassNameContract } from "@microsoft/fast-components-class-name-contracts-msft";
 import manageJss, { ManagedJSSProps } from "@microsoft/fast-jss-manager-react";
 import { DesignSystem, SubheadingStyles } from "@microsoft/fast-components-styles-msft";
-import { Subtract } from "utility-types";
 import subheadingSchema from "./subheading.schema";
 import subheadingSchema2 from "./subheading.schema.2";
 import MSFTSubheading, {
@@ -22,9 +21,9 @@ import MSFTSubheading, {
 const Subheading = manageJss(SubheadingStyles)(MSFTSubheading);
 type Subheading = InstanceType<typeof MSFTSubheading>;
 
-type SubheadingHandledProps = Subtract<
+type SubheadingHandledProps = Omit<
     MSFTSubheadingHandledProps,
-    SubheadingManagedClasses
+    keyof SubheadingManagedClasses
 >;
 type SubheadingProps = ManagedJSSProps<
     MSFTSubheadingProps,