+++ packages/react/fast-layouts-react/src/page/index.ts
@@ -1,6 +1,5 @@
 import React from "react";
 import manageJss, { ManagedJSSProps } from "@microsoft/fast-jss-manager-react";
-import { Subtract } from "utility-types";
 import {
     Page as BasePage,
     PageHandledProps as BasePageHandledProps,
@@ -13,7 +12,7 @@ import {
 const Page = manageJss()(BasePage);
 type Page = typeof Page;
 
-type PageHandledProps = Subtract<BasePageHandledProps, PageManagedClasses>;
+type PageHandledProps = Omit<BasePageHandledProps, keyof PageManagedClasses>;
 type PageProps = ManagedJSSProps<BasePageProps, PageClassNamesContract, undefined>;
 
 export { Page, PageProps, PageHandledProps, PageUnhandledProps, PageClassNamesContract };